requirements.txt
insta-wizard>=0.1.7
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
rich>=13.0
